Some warning fixes.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@31458 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/win32/hs-isa.c
@@ -157,7 +157,7 @@ static BYTE hardsid_inb(unsigned int addrint)
             winio_inp32fp(addr, &tmp, 1);
             retval = (BYTE)tmp;
         } else {
-            retval = inpout_inp32fp(addr);
+            retval = (BYTE)inpout_inp32fp(addr);
         }
     } else {
 #ifdef  _M_IX86
